refactor: extract capture helpers to _capture module
The Playwright-based screenshot and GIF capture functions had grown to
dominate _cli.py. Moving them to _capture.py makes the boundary between
CLI command wiring and browser-capture logic explicit in the code structure
itself, removing the need for section-separator comment banners.
